Remove old login-before-roster UI code
