--- src/generic_ui/polymer/accept-user-invite.html
@@ -1,64 +1,0 @@
-<link rel='import' href='../../bower/core-icons/core-icons.html'>
-<link rel='import' href='../../bower/core-overlay/core-overlay.html'>
-<link rel="import" href="../../bower/core-signals/core-signals.html">
-<link rel='import' href='../../bower/paper-input/paper-input-decorator.html'>
-<link rel='import' href='../../bower/polymer/polymer.html'>
-<link rel='import' href='button.html'>
-<link rel="import" href="i18n-filter.html">
-
-<polymer-element name='uproxy-accept-user-invite'>
-  <template>
-    <style>
-      #acceptUserInvitePanel {
-        height: 100%;
-        width: 100%;
-        background-color: white;
-      }
-      .section {
-        padding-left: 2em;
-        padding-right: 2em;
-        padding-top: 1em;
-        padding-bottom: 1em;
-        line-height: 20px;
-      }
-      a {
-        cursor: pointer;
-      }
-      uproxy-button {
-        margin: 0;
-        padding-left: 1em;
-        padding-right: 1em;
-      }
-      paper-input-decorator {
-        font-size: 16px;
-      }
-    </style>
-
-    <core-signals on-core-signal-open-accept-user-invite-dialog='{{ openAcceptUserInvitePanel }}'></core-signals>
-
-    <core-overlay id='acceptUserInvitePanel'>
-      <uproxy-app-bar on-go-back='{{ closeAcceptUserInvitePanel }}'>
-        {{ "ENTER_INVITATION" | $$ }}
-      </uproxy-app-bar>
-
-      <div class='section'>
-        <p>{{ "ENTER_INVITATION_DESCRIPTION" | $$ }}</p>
-
-        <div id='addFriend'>
-          <paper-input-decorator label="{{ 'ENTER_INVITATION' | $$ }}">
-            <input is='core-input' value='{{ receivedInviteToken }}'>
-          </paper-input-decorator>
-          <uproxy-button affirmative autofocus on-tap='{{ acceptInvitation }}'>
-            {{ 'SUBMIT_BUTTON_TEXT' | $$ }}
-          </uproxy-button>
-        </div>
-      </div>
-
-      <div class='section'>
-        <a on-tap='{{ showInviteUser }}'>{{ 'INVITE_LINK_FROM_ACCEPT_PAGE' | $$ }}</a>
-      </div>
-
-    </core-overlay>
-  </template>
-  <script src='accept-user-invite.js'></script>
-</polymer-element>
