--- src/generic_ui/polymer/network.html
@@ -1,42 +1,0 @@
-<link rel="import" href="../../bower/polymer/polymer.html">
-<link rel='import' href='button.html'>
-
-<polymer-element name='uproxy-network' attributes='networkName'>
-  <template>
-    <style>
-      :host {
-        display: block;
-        box-sizing: border-box;
-      }
-      uproxy-button {
-        width: 200px !important;
-      }
-      uproxy-button.GMail {
-        background: #dd4b39;
-      }
-      uproxy-button.Facebook,
-      uproxy-button.Facebook-Firebase-V2 {
-        background: #3b5998;
-      }
-      uproxy-button.WeChat {
-        background: #6ab000;
-      }
-      .experimental {
-        font-style: italic;
-        font-size: 12px;
-      }
-    </style>
-
-    <div class='network'>
-      <span>
-        <span>
-          <uproxy-button raised on-tap='{{connect}}' class='{{ networkName }}'>
-            {{ displayName }}
-          </uproxy-button>
-          <div class='experimental' hidden?='{{ !isExperimental }}'>{{ "EXPERIMENTAL" | $$}}</div>
-        </span>
-      </span>
-    </div>
-  </template>
-  <script src='network.js'></script>
-</polymer-element>
