+++ src/generic_ui/polymer/root.html
@@ -10,7 +10,6 @@
 <link rel="import" href="../../bower/core-tooltip/core-tooltip.html">
 <link rel="import" href="../../bower/core-overlay/core-overlay.html">
 <link rel='import' href='invite-user.html'>
-<link rel='import' href='accept-user-invite.html'>
 <link rel="import" href="i18n-filter.html">
 <link rel="import" href='splash.html'>
 <link rel="import" href='faq.html'>
@@ -284,7 +283,6 @@
     <!-- core-overlays, launched using core-signals -->
     <uproxy-invite-user></uproxy-invite-user>
     <uproxy-google-invite></uproxy-google-invite>
-    <uproxy-accept-user-invite></uproxy-accept-user-invite>
     <uproxy-feedback id='feedback'></uproxy-feedback>
     <uproxy-faq id='faq'></uproxy-faq>
     <uproxy-advanced-settings id='advancedSettings'></uproxy-advanced-settings>
