+++ src/generic_ui/polymer/settings.html
@@ -211,7 +211,7 @@
       <div id='accountChooser' class='settingsContent' hidden?='{{ !accountChooserOpen }}'>
         <span class='smalltext'>{{ "CONNECTED_ACCOUNTS" | $$ }}</span>
         <template repeat='{{ name in model.networkNames }}'>
-          <div hidden?="{{name=='Cloud' && !model.globalSettings.showCloud && !ui.showRosterBeforeLogin}}">
+          <div hidden?="{{name=='Cloud' && !model.globalSettings.showCloud}}">
             <uproxy-network-in-settings name='{{ name }}'></uproxy-network-in-settings>
           </div>
         </template>
